DDC Http Store - Added ability to configure AuthScheme (defaulting to JWTBearer as before) and the ability to directly specify a auth token. This can be used to run with the ServiceAccount scheme in Cloud DDC which we intend to use for getting onboarded quicker by licensees (it is not a recommended approch to use for wider deployments across teams).
#rb Zousar.Shaker

Restore the printing of the response body when an unexpected error is received, but now with consideration for the content type of the response.

#rb Devin.Doucette
#okfogithub public
Change the strategy for HttpCacheStore Put operations to have larger and distinct queues for each of the put operation phases.  Instead of one queue of 24 maximum in-flight requests, we can now have 64 PutRef requests AND 64 PutBlobs  requests, and 64 PutFinalize requests operating simultaneously.  The purpose is to allow per-shader caching to:

1. Put a higher volume of individual cache items in a lesser amount of wall time
2. Reduce the gap in time between when a ref is put and when it is finalized during times of heavy Put workload

Along with this change there is a System.DerivedDataCache.HttpDerivedDataBackend.CacheStoreStressPut automated test (StressFilter) that puts 1000 records containing 12 byte values in each.  Before this change, the test took 21 seconds to complete.  It now takes 9 seconds to complete.  There are opportunities to improve this further through batching.

#rb Devin.Doucette

Don't attempt immediate retry when a 502 error happens during communication with the UE Cloud DDC server.

#[fyi] joakim.lindqvist

#ushell-cherrypick of 29716222 by Joakim.Lindqvist
#ushell-cherrypick of 29729465 by Zousar.Shaker
#ushell-cherrypick of 29913181 by Zousar.Shaker
#ushell-cherrypick of 30024810 by Zousar.Shaker

[CL 30044720 by zousar shaker in ue5-main branch]