Fixed inconsistent tooltip for Blueprint config variables.
GetConfigName doesn't always return the base name of an .ini file, which is what we want. For example, if the class uses DefaultEngine.ini for its config values, GetConfigName will return the base name, "Engine". However, if the class uses DefaultEditor.ini, GetConfigName returns the path to the merged .ini file that's located under the game's "Saved/Config" directory.

Since ClassConfigName corresponds to the base name of a config file, we can just use that for the tooltip.

#rb ben.hoffman

[CL 30137097 by dave jones2 in ue5-main branch]