[PCG] Multi-line node titles
Nodes have an authored title line (primary) and a generated title line (secondary). Primary line tends to have most important/relevant information and can depend on whether user has authored a title or not.

Users requested some nodes have flipped order so that the generated part is displayed first, helps for e.g. BP nodes where the key bit of info is the executed blueprint.

Text clipping/overflow flipped for generated title lines so that the end of the line is always visible which usually has the key info.

#jira UE-201565
#rb julien.lheureux, wyatt.marvil
[FYI] JeanSebastien.Guay, antoine.guilbaud

[CL 30158846 by huw bowles in ue5-main branch]