Natvis: Implemented a workaround for 10 second freezes when debugging on some platforms (caused by CustomListItems visualizer in TSet, in particular).
On the affected platforms the Size parameter of CustomListItems is incorrectly handled. Instead of breaking the loop iteration after the given number of items have been added, the traversal continues (up to MaxItemsPerView), it just sort of divides the results into batches having Size items. This iteration is expensive as we keep dereferencing invalid container memory using in <If Condition="..."> attribute. As a result, we had these freezes even for empty containers.

Resolved by adding an explicit stop condition for the loop. The problem was discovered for TSet but I've updated all visualizers using CustomListItems to be sure we don't run into the same problem elsewhere.

It's been reported to the 1st party. Once a fix is in place, we willl be able to remove my workaround if we choose so, but there is really nothing wrong in having the stop condition explicitly stated.

#rb robert.millar
#rnx

[CL 30170687 by wojciech krywult in ue5-main branch]