PR #10925: Fix UserWidget's PlayAnimationForward and PlayAnimationReverse not playing animations if the animation was recently stopped
In PlayAnimationForward/PlayAnimationReverse, just return the player if it's already playing. GetSequencePlayers should only be returning players that are NOT stopped.  However, there is the possibility that Stop() has been called on a player, but its status has not been changed to Stopped. In that case, this check will be bypassed and PlayAnimation will be called so that the animation will play.

#rb Andrew.Rodham
#jira UE-196655

[CL 30187280 by DoubleDeez in ue5-main branch]