UGS: Extended the "SDK Info" window to support adding arbitrary badge buttons.
Previously, this dialog only had fixed support for Install and Browse commands. While sufficient in most cases, on some platforms they lacked a bit of additional flexibility.

Now, we may add badges with custom names launching arbitrary programs with optional command-line parameters. For instance,

    ReleaseNotes=(Program="notepad.exe", Args="notes.txt")

Note that now actions are added in the order in which are defined in the .ini config.

The underlying implementation has been unified so Install, Browse and the new actions launch programs in exactly the same way, but the previous behavior and syntax for Install and Browse is still supported.

#rb David.Harvey, randy.jones
#rnx

[CL 30217662 by wojciech krywult in ue5-main branch]