Account for complete event fitting in the buffer for scoped events.
If memory tracing is enabled we need to make sure that both the enter uid and the event header fits into the current buffer. Otherwise it's possible that while requesting a new buffer and one has to be allocated, allocation trace events are inserted between the enter uid and the event. The result is that allocation event is incorrectly marked as a scope.

#rb martin.ridgers, ionut.matasaru
#jira UE-200495

[CL 30259854 by johan berg in ue5-main branch]