[PCG] Fix debug object selection getting lost when moving debug flag around.
Only force removal of PCG components if higen grid sizes change. We added a new change type to communicate this.

To make this work, we also made the runtime gen system support a more refresh-like path - previously a refresh would completely clear out the PAs and recreate them. Now the PAs are marked dirty and scheduled for generation which is much nicer.

Also ensure PA grid updated when changing & undoing Grid Size changes, which required adding some code here and there to the refresh paths that respond to the GenerationGrid change type.

#rb julien.lheureux, wyatt.marvil

[CL 30316108 by huw bowles in ue5-main branch]