Add directory watchers for plugin root directories in Asset Registry
The directory watchers in the asset registry take up just above 100MB of memory when running on Windows. We are trying to reduce the memory footprint of the Content Worker to mitigate OOM scenarios. The main reason for this is the sheer amount of directory watches that we are setting up (around 418, as of writing). Most of these share a common prefix (either Engine/Plugins or <Game>/Plugin), so we may instead set up watchers for these directories and filter all incoming changes appropriately. This results in only 10 directory watchers, bringing memory usage from 104.5MB to just about 2.5MB.

#rb Matt.Peters, rob.perren

[CL 30316166 by sebastian schoner in ue5-main branch]