Check whether plugin directories exist before adding a directory watcher to them
Those directories might not exist, in which case the directory watcher issues a warning. There is no point in watching these directories if they do not exist as there likely won't be any mount points below them (that is different from the per-plugin paths because there the plugin might exist and it's just missing the last part of the path). In a worst case scenario, the session will set up a new directory watcher if a new mount point appears at some point during the lifetime of the program.

[FYI] matt.peters

[CL 30317760 by sebastian schoner in ue5-main branch]