Sequencer: Corrected material parameters not removing hierarchical blend targets correctly
When a material parameter was previously subject to a hierarchical blend target continues to be animated but without the blend target, it was previously just removing the blend target component. This has the side effect of the hierarchical blend channel remaining on the entity, but it will not be visited by FinalizeBlendTargets. Changing this code to match the property instantiator by adding the RemoveHierarchicalBlendTarget tag ensures it gets tidied up correctly while maintaining the ordering dependencies:

hierarchical easing instantiator -> material parameter system -> hierarchical easing finalization system

This code should have been updated as part of CL 28275348, but was evidently overlooked

#rb ludovic.chabant

[CL 30318595 by andrew rodham in ue5-main branch]