Fix incorrect Receipt.Launch value if there are executable-type runtime dependencies.
The bug was introduced with 17546102. Previously "Launch" would get set to the first suitable BuildProduct entry; the module's binary. But after that change it can get incorrectly set to the last suitable BuildProduct entry if the target doesn't not produce a "-Cmd" suffixed output. Any module (such as Sentry's UE integration [1]) that introduces a RuntimeDependency on .exe files can result in invalid target receipts.

[1] https://github.com/getsentry/sentry-unreal/blob/main/plugin-dev/Source/Sentry/Sentry.Build.cs

#rb Joe.Kirchoff
#rnx

[CL 30408745 by martin ridgers in ue5-main branch]