Path Tracing: Simplify logic in TraceTransparentRay to reduce live state, simplify decal math
The main change here is making the decal logic accumulate inside the decal shader itself so we can reduce the amount of logic in the path tracer and also reduce the amount of live state required for the decal calls to TraceRay (532 bytes down to 456 bytes) and CallShader (544 bytes down to 468 bytes). This appears to improve performance on decal heavy scenes.

Performance is slightly reduced in other scenes which do not use decals. This can probably be compensated for by further working to reduce the live state by applying similar optimizations to the other payloads.

#jira UE-202786

[CL 30414146 by chris kulla in ue5-main branch]