[NNE] Add opset versioning to NNEQA Problem Versioned operators require proper testing infrastructure.
Solution
Complement NNEQA json test description format with syntax to express information regarding which versions of the operators to test and how.
Introduce a new type of tests (opset tests) for testing all the supported versions of the ops. Contrarily to iteration tests (those we had so far, testing a single version) they are not supposed to be run on every development iteration, but with a much lower frequency, due to their length.
Through the newly introduced `default_version` and `versions` fields of a test target it's possible to, respectively, indicate which version of the op should be tested in iteration tests and which versions are supported overall (tested in opset tests). Both are optional fields.
`versions` is also a field of input-output sets and attribute sets, restricting the pool of versions the corresponding tests apply to.

Note: opset tests have been added only for a few OPs here. Another CL should add opset tests for the remaining OPs.

Tests
- NNEIncubatorEditor Win64
- Non-unity compile

#rnx
#rb florent.guinier
#jira UE-201161

[CL 30427746 by giacomo serafini in ue5-main branch]