Follow-up fix for CL 30398348: don't set EditInline on actor components.
CL 30398348 introduced a loophole for actor component instancing. By setting a component variable to "EditInline", it was now possible to instance a component in both the Details panel *and* the Components panel. Additionally, instancing in the Details panel would actually outer the component to the CDO instead of the UClass.

Note that component variables will default to using SPropertyEditorAsset in the Details panel, which is also misleading, since it implies that a non-null default value can be set on the component property. UE-203267 was created to improve the messaging here.

#rb Phillip.Kavan

[CL 30434282 by dave jones2 in ue5-main branch]