Print out shader compile worker memory usage when main process goes OOM
When we encounter an OOM-crash in UEFN's cloud cooking, we would like to be able to quickly rule out different parts of the cooking process. One major part that we have poor visibility on at the moment is shader compilation. This CL adds a codepath that prints out the memory usage of the shader compile processes when such a crash occurs.

#rb Laura.Hermanns

[CL 30479044 by sebastian schoner in ue5-main branch]