Print out total shader compile worker memory usage on OOM
Usually the first thing I do when I see the shader compile worker memory report is to sum it up. We might as well just add the sum to the report. Summing up maximum values of course isn't going to reflect the actual maximum, but it does provide an upper bound for the maximum memory usage over the lifetime of the shader compile workers, which can be useful to rule out shader compilation as a source of memory issues.

#rb Laura.Hermanns

[CL 30480208 by sebastian schoner in ue5-main branch]