Sequencer: change how the sequence observer is queried to solve networking bugs
The observer is now checked at all public API entry points for whether a sequence can be observed or not. Before, it would only be checked when updating the sequence, i.e. a non-observable sequence would still actually be playing... it would just not be updating. Now, a non-observable sequence is not playing at all.

This has the effect of letting us try again in PostNetReceive if needed. If the server started playing and the client wants to start playing too, it will now check if the sequence is observable.

1. If it is observable, it starts playing.

2. If it isn't observable, it doesn't start playing, but on the next PostNetReceive it will try again. For a truly non-observable sequence, this means trying again all the time and always refusing to. This isn't any worse than before, when the sequence was playing but always refusing to update every frame. In fact, it's a bit less worse, since PostNetReceive doesn't necessarily happen every frame.

3. If it is observable, but network replication of instigator properties hasn't happened yet, then the observer mistakenly tells the client that the sequence isn't observable, and nothing happens... but eventually, network replication does come through, and the client finally starts playing, with a delay.

#rb Max.Chen, Andrew.Rodham

[CL 30498852 by ludovic chabant in ue5-main branch]