UE-203565 - OnBrowseSocket BP Menu failing to open when the SubobjectEditor list is docked to the side panel or closed.
When a window is docked in the Blueprint editor, it gets added to the STabSidebar, which is created by the toolkit host. If we use the current SSubobjectEditor as the widget path, we won't actually find a window, because it now lives under the STabSidebar. As a result, the drop down menu for mesh sockets would fail to open.

The easiest fix is to simply use the toolkit host's parent widget, which finds the correct widget path in the docked and undocked cases.

#jira UE-203565
#rb ben.hoffman

[CL 30538362 by dave jones2 in ue5-main branch]