Fixed a race condition in TEDS' Scratch Buffer.
The call to `GetEmptyBlock()` no longer uses an owner index, but instead uses a more traditional implementation of lock-free queue. The original implementation tried to avoid the ABA problem but since later updates to the scratch buffer started using seperate queues for available and recycled blocks, the ABA problem it tried to solve can no longer happen. A simpler implementation is therefore warranted and also fixes the deadlock.
The following changes are made in addition:
- It's now possible to have multiple instances of the scratch buffer. This has mostly been added to supporte testing.
- The scratch buffer is now stored in a shared pointer. This was needed to allow multiple instances of the scratch buffer to exist. Some optimization opportunities remain.
- Tests were added to cover basic functionality and do multi-threaded stress testing.

#jira UE-203667
#rb logan.buchy

[CL 30572315 by ronald koppers in ue5-main branch]