fix backbuffer resize: we were trying to release the swapchain buffers before their UAV was actually released.
Moving the UAV deletion code just above doesn't work though: FD3D12Viewport::Resize is called from the game thread. When destroyed, Backbuffer UAVs are pushed to the RHI's PendingDeletes queue before being actually released.

Because of that, we also need to ensure the queue is processed before trying to destroy the swapchain buffers. ImmediateFlush(EImmediateFlushType::FlushRHIThreadFlushResources) takes care of that, but FlushRenderingCommands() already does it + add the GameThread-RenderThread synchronization

#rb Luke.Thatcher
#jira UE-203473

[CL 30630793 by benjamin rouveyrol in ue5-main branch]