[mutable] Exposed Bake instance method by extracting it from "CustomizableObjectEditorViewportClient" - This will make possible the invocation of the baking externally to the viewport client. - Replaced UE with Mutable notifications (prevents log spamming) - Added Multicast delegate for mutable instance update with result - Other small changes
Note : part of the chnages in this CL does not make sense without a context. The main body of the baking operation has been placed where it was originally (with some changes) to ease the review process. In a later CL, when the baking test commandled is implemented, the code will be placed where it should.

[REVIEW] [at]gerard.martin [at]pere.rifa
#rb gerard.martin, pere.rifa

[CL 30680260 by daniel moreno in ue5-main branch]