Path Tracer: Avoid pack/unpack of payload in the main trace loop
In the loop over transparent hits, unpacking the payload immediately increases the live state substantially. After this change, the live state for the main TraceRay calls reduces from 404 bytes down to 282 bytes. Additionally, some of the logic around decals can be simplified slightly.

The observed speedup is around 5-10% depending on the scene with simpler scenes seeing a larger gain.

#jira UE-202786

[CL 30693663 by chris kulla in ue5-main branch]