Fix crashes around creating and deleting pose assets
Found an additional crash when investigating this ticket - when previewing a newly-created (i.e. unsaved) pose asset, moving the pose preview sliders would crash. Adding a refresh of the curve index table to fix this.
Original crash was caused by calling into the anim graph when the skeletal mesh is null. No nodes are built to be robust to this, so adding an guard similar to that in USkeletalMeshComponent::InitAnim()

#jira UE-198567
#rb Nicholas.Frechette

[CL 30715629 by thomas sarkanen in ue5-main branch]