Fixed type information columns in TEDS going stale in some cases.
UObject based references in TEDS also store a reference to their type information. If that type information is pointing to script data the type information can go stale after a script compile. This is something that can happen in general with type information, scripts are just the first use case where this was discovered in practice. The solution is to list all the type information pointers in TEDS when a request to list dependencies comes in. This way the type information gets refreshed and is valid again. The approach to refreshing type information through `AddReferencedObjects` caused issues with the garbage collection while the version that used the `OnObjectsReinstanced` version caused order-of-execution errors as the updates wouldn't be processed before objects were reinstanced. The latter solution was still preferred as it's a more targeted solution. To address the order-of-execution issues, the type-to-table lookup is delayed until the objects are processed and added to the data storage.
This change also includes an update to some of the debug commands so they're in a central place, plus some new commands were added.

#jira UE-201532
#rb aditya.ravichandran, logan.buchy

[CL 30851293 by ronald koppers in ue5-main branch]