MeshElementSelection: Make selectable mesh elements visible.
When we have a valid topology type (Triangle or Poly, not None), we render selectable mesh elements.

Added AccumulateElementsFromPredicate to the GeometrySelector interface which allows a predicate function to filter which geometry elements ought to be included in the provided FGeometrySelectionElements. The predicate is aware of the type of element (vert, edge, tri) and the selected element ID (which is a single value encoding the topology ID and the element ID). The predicate returns true for elements to include and false for elements to exclude. Internally, this function builds transient vert, edge, and tri selections with this predicate and then accumulates all selection elements into the provided FGeometrySelectionElements.

#rb lonnie.li
#jira UE-202674

[CL 30862050 by nickolas drake in 5.4 branch]