Sequencer: fix splitting/trimming of looping sub-sections
Splitting a section is, by default, done by cloning the section, trimming the original on the right, and trimming the clone on the left.
Trimming looping sub-sections on the right is straight-forward. Trimming them on the left requires adjusting the FirstLoopStartFrameOffset to preserve where the loops happen. This was previously done directly in the SplitSection method, which meant that it was roughly OK for splitting, but that trimming in-and-of-itself was broken. This was later fixed, but the code in SplitSection remained, which meant that we were fixing up FirstLoopStartFrameOffset twice. This change therefore just removes the code in SplitSection.

In addition to that, the code in TrimSection had a bug: the StartFrameOffset property is already defined in the inner sub-sequence's time space, so there is no need to transform it using the OuterToInnerTransform function.

#jira UE-204176
#rb david.bromberg, Max.Chen

[CL 30864739 by ludovic chabant in ue5-main branch]