Rewrite of TEDS' Scratch Buffer.
This change addresses a memory stomp that happened, which was caused by an incorrect use of placement new.
Originally the issue was thought to be with the Scratch Buffer's custom allocator so the choice was made to move to FConcurrentLinearBulkObjectAllocator. This has been kept in to reduce future maintenance work. The implementation of `AddColumnUnitialized` has also been updated to remove the near duplicate implemenation.

#jira UE-203667
#rb logan.buchy

[CL 30905729 by ronald koppers in ue5-main branch]