New method for allowing particles spawned from NDC to access the NDC element that generated them.
Work using an indirection table that allocates out power of 2 chunks of instances to each NDC item according to how many particles each spawns.
This is a bit convoluted but provides a good balance of minimal size of the table and allowing both high particle counts per NDC and high NDC counts spawning few particles.
Method is described in greater detail in code comments.

New function GetNDCSpawnData(). Returns the NDC index for a particle as well as the total count for that NDC and the spawn index for the current particle inside that NDC spawn. i.e NDC local exec index.

Previous SpawnGroup based method still works as a backup where needed.
There may be cases where the new method cannot work. For example if a proper ExecIndex per NDC is needed.

Misc:
- Fixed bug in translator that meant emitter.ID would not be set for GPU scripts.
#rb rob.krajcarski
#jira UE-195800

[CL 30956725 by simon tovey in 5.4 branch]