MovieGraph: Add support for ScriptExecutionNodes which can run code during a render. To use this feature, implement a class that inherits from UMovieGraphScriptBase (in either Python or C++), and then add a UMovieGraphExecuteScriptNode node to your Graph's configuration. Within the Script Node, specify the script that you want run.
At runtime, Movie Render Graph will create an instance of your script before the job starts. It will then persist this instance throughout the job until the end, at which point the instance will be destroyed. During the runtime, various callbacks will be called on your script and it will be provided data from the current graph.

The graph duplicates the Job, and Graph Configuration Assets before passing them to the scripting environment. This allows your script to make changes to these settings (such as overriding a variable on the job, or adding nodes to the graph) without having to worry about restoring these changes at the end of the job.

#jira UE-200132, UE-202976, UE-205154
#rb chad.williams

[CL 30983441 by matt hoffman in ue5-main branch]